LayoutTests/http/tests/security/xssAuditor/formaction-on-button-expected.txt

CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

ALERT: formaction present on BUTTON with value of about:blank
